requirements.txt
bidict==0.21.4
cachelib==0.6.0
click==8.0.3
Deprecated==1.2.13
dnspython==2.2.0
eventlet==0.33.0
Flask==2.0.2
Flask-Session==0.4.0
Flask-SocketIO==5.1.1
greenlet==1.1.2
h11==0.13.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pyparsing==3.0.7
pyserial==3.5
python-engineio==4.3.1
python-socketio==5.5.1
simple-websocket==0.5.0
six==1.16.0
Werkzeug==2.0.2
wrapt==1.13.3
wsproto==1.0.0
zope.event==4.5.0
zope.interface==5.4.0
Adafruit-Blinka==7.1.0
adafruit-circuitpython-dht==3.7.1
RPi.GPIO==0.7.1
